manually generate knock css vars
