+++ packages/react/vite.config.mts
@@ -6,8 +6,6 @@ import { LibraryFormats, defineConfig, loadEnv } from "vite";
 import dts from "vite-plugin-dts";
 import noBundlePlugin from "vite-plugin-no-bundle";
 
-import fs from "fs/promises";
-
 // https://vitejs.dev/config/
 export default defineConfig(({ mode }) => {
   const env = loadEnv(mode, process.cwd(), "");
@@ -58,6 +56,8 @@ export default defineConfig(({ mode }) => {
         },
         plugins: [
           execute([
+            // Move index.css to root of dist
+            `mv dist/esm/index.css dist/index.css`,
             // Delete extra .css.js files
             `find ./dist -name "*.css.js" -delete`,
             `find ./dist -name "*.css.js.map" -delete`,
@@ -77,53 +77,6 @@ export default defineConfig(({ mode }) => {
                     /(import ".*?\.css\..*?";)|(require\(['"][^()]+\.css(\.js)?['"]\);?)/g;
                   file.code = file.code.replace(pattern, "");
                 }
-              }
-            },
-          },
-          {
-            name: "generate-css-tokens",
-            enforce: "post",
-            closeBundle: async () => {
-              try {
-                // 1. Update Telegraph tokens
-                const tokensFile = path.resolve(
-                  __dirname,
-                  "../../node_modules/@telegraph/tokens/dist/css/default.css",
-                );
-
-                let tokensCss = await fs.readFile(tokensFile, "utf-8");
-                // a. Use --knock css variable prefix
-                tokensCss = tokensCss.replace(/--tgph/g, "--knock");
-                // b. Update color mode selector
-                tokensCss = tokensCss.replace(
-                  /data-tgph-appearance/g,
-                  "data-knock-color-mode",
-                );
-                // c. Rename `accent` color to `orange`
-                tokensCss = tokensCss.replace(
-                  /--knock-accent/g,
-                  "--knock-orange",
-                );
-
-                // 2. Add to index.css file
-                let indexCss = await fs.readFile(
-                  path.resolve(__dirname, `dist/${buildTarget}/index.css`),
-                  "utf-8",
-                );
-                indexCss = tokensCss + indexCss;
-
-                // 3. Write index.css to root of dist
-                await fs.writeFile(
-                  path.resolve(__dirname, "dist/index.css"),
-                  indexCss,
-                );
-
-                // 4. Delete build target index.css
-                await fs.rm(
-                  path.resolve(__dirname, `dist/${buildTarget}/index.css`),
-                );
-              } catch (error) {
-                console.error("Error processing CSS:", error);
               }
             },
           },
